# laba1/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "API сервер приложений",
      "description": "Трёхзвенная архитектура: клиент ↔ сервер приложений (этот API) ↔ PostgreSQL. Аутентификация по JWT. Роли: **admin** (полный доступ), **moderator** (создание/редактирование), **viewer** (только чтение).",
      "version": "0.1.0"
    },
    "paths": {
      "/api/auth/register": {
        "post": {
          "tags": [
            "Аутентификация"
          ],
          "summary": "Регистрация пользователя",
          "operationId": "register_api_auth_register_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/auth/login": {
        "post": {
          "tags": [
            "Аутентификация"
          ],
          "summary": "Вход",
          "operationId": "login_api_auth_login_post",
          "requestBody": {
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Body_login_api_auth_login_post"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Token"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/auth/me": {
        "get": {
          "tags": [
            "Аутентификация"
          ],
          "summary": "Текущий пользователь",
          "operationId": "me_api_auth_me_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/api/items": {
        "get": {
          "tags": [
            "Сущности"
          ],
          "summary": "Список сущностей",
          "operationId": "list_items_api_items_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "offset",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 0,
                "default": 0,
                "title": "Offset"
              }
            },
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 1,
                "default": 20,
                "title": "Limit"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ItemResponse"
                    },
                    "title": "Response List Items Api Items Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Сущности"
          ],
          "summary": "Создать сущность",
          "operationId": "create_item_api_items_post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemCreate"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ItemResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/items/{item_id}": {
        "get": {
          "tags": [
            "Сущности"
          ],
          "summary": "Получить сущность по ID",
          "operationId": "get_item_api_items__item_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "item_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Item Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ItemResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "patch": {
          "tags": [
            "Сущности"
          ],
          "summary": "Обновить сущность",
          "operationId": "update_item_api_items__item_id__patch",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "item_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Item Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ItemResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Сущности"
          ],
          "summary": "Удалить сущность",
          "operationId": "delete_item_api_items__item_id__delete",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "item_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Item Id"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful Response"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/health": {
        "get": {
          "tags": [
            "Служебные"
          ],
          "summary": "Health",
          "operationId": "health_health_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Body_login_api_auth_login_post": {
          "properties": {
            "grant_type": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^password$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grant Type"
            },
            "username": {
              "type": "string",
              "title": "Username"
            },
            "password": {
              "type": "string",
              "format": "password",
              "title": "Password"
            },
            "scope": {
              "type": "string",
              "title": "Scope",
              "default": ""
            },
            "client_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            },
            "client_secret": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "format": "password",
              "title": "Client Secret"
            }
          },
          "type": "object",
          "required": [
            "username",
            "password"
          ],
          "title": "Body_login_api_auth_login_post"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "ItemCreate": {
          "properties": {
            "title": {
              "type": "string",
              "maxLength": 256,
              "minLength": 1,
              "title": "Title"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 1024
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          },
          "type": "object",
          "required": [
            "title"
          ],
          "title": "ItemCreate"
        },
        "ItemResponse": {
          "properties": {
            "title": {
              "type": "string",
              "maxLength": 256,
              "minLength": 1,
              "title": "Title"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 1024
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "title",
            "id"
          ],
          "title": "ItemResponse"
        },
        "ItemUpdate": {
          "properties": {
            "title": {
              "type": "string",
              "maxLength": 256,
              "minLength": 1,
              "title": "Title"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 1024
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          },
          "type": "object",
          "required": [
            "title"
          ],
          "title": "ItemUpdate"
        },
        "Role": {
          "type": "string",
          "enum": [
            "admin",
            "moderator",
            "viewer"
          ],
          "title": "Role"
        },
        "Token": {
          "properties": {
            "access_token": {
              "type": "string",
              "title": "Access Token"
            },
            "token_type": {
              "type": "string",
              "title": "Token Type",
              "default": "bearer"
            },
            "exp": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exp"
            }
          },
          "type": "object",
          "required": [
            "access_token"
          ],
          "title": "Token"
        },
        "UserCreate": {
          "properties": {
            "username": {
              "type": "string",
              "maxLength": 64,
              "minLength": 2,
              "title": "Username"
            },
            "email": {
              "type": "string",
              "format": "email",
              "title": "Email"
            },
            "password": {
              "type": "string",
              "minLength": 6,
              "title": "Password"
            },
            "role": {
              "$ref": "#/components/schemas/Role",
              "default": "viewer"
            }
          },
          "type": "object",
          "required": [
            "username",
            "email",
            "password"
          ],
          "title": "UserCreate"
        },
        "UserResponse": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "username": {
              "type": "string",
              "title": "Username"
            },
            "email": {
              "type": "string",
              "format": "email",
              "title": "Email"
            },
            "role": {
              "$ref": "#/components/schemas/Role"
            },
            "is_active": {
              "type": "boolean",
              "title": "Is Active"
            }
          },
          "type": "object",
          "required": [
            "id",
            "username",
            "email",
            "role",
            "is_active"
          ],
          "title": "UserResponse"
        },
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            },
            "input": {
              "title": "Input"
            },
            "ctx": {
              "type": "object",
              "title": "Context"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {
  
              },
              "tokenUrl": "/api/auth/login"
            }
          }
        }
      }
    }
  }